Refactor URL builders in UI
