--- ui/src/utils/baseUrl.js
@@ -1,8 +1,0 @@
-import config from '../config'
-
-export const baseUrl = (path) => {
-  const base = config.baseURL || ''
-  const parts = [base]
-  parts.push(path.replace(/^\//, ''))
-  return parts.join('/')
-}
