--- ui/src/utils/docsUrl.js
@@ -1,1 +1,0 @@
-export const docsUrl = (path) => `https://www.navidrome.org${path}`
