+++ ui/src/utils/index.js
@@ -1,7 +1,5 @@
-export * from './baseUrl'
-export * from './docsUrl'
 export * from './formatters'
 export * from './intersperse'
 export * from './notifications'
 export * from './openInNewTab'
-export * from './shareUrl'
+export * from './urls'
